Migrate serialization tests to native protocol + prepare arrays support
