+++ integration-testing/IntegrationTests.hs
@@ -1,23 +1,36 @@
 {-# LANGUAGE
-  OverloadedStrings
+    OverloadedStrings
+  , TypeApplications
 #-}
 
 module IntegrationTests
   ( main
   ) where
 
-import ClickHaskell.Client (ChCredential(..))
-import IntegrationTests.Serialization (runSerializationTests)
+-- Internal
+import ClickHaskell (ChCredential(..), openNativeConnection)
+import ClickHaskell.DbTypes
+import IntegrationTests.Serialization (runSerializationTest)
 import IntegrationTests.WriteReadEquality (runWriteReadEqualityTest)
+
+-- GHC included
+import Data.ByteString as BS (singleton)
 
 main :: IO ()
 main = do
-  let cred =
-        MkChCredential
-          { chLogin="default"
-          , chPass=""
-          , chUrl="http://localhost:8123"
-          , chDatabase="default"
-          }
-  runSerializationTests cred
-  runWriteReadEqualityTest cred
+  let credentials = MkChCredential
+        { chLogin    = "default"
+        , chPass     = ""
+        , chHost     = "localhost"
+        , chDatabase = "default"
+        , chPort     = "9000"
+        }
+  connection <- openNativeConnection credentials
+  
+  runSerializationTest @ChInt32 connection [minBound, toEnum 0, maxBound]
+  runSerializationTest @ChInt64 connection [minBound, toEnum 0, maxBound]
+  runSerializationTest @ChUInt32 connection [minBound, toEnum 0, maxBound]
+  runSerializationTest @ChUInt64 connection [minBound, toEnum 0, maxBound]
+  runSerializationTest @ChString connection (map (toChType . BS.singleton) [1..255])
+  -- runSerializationTest @(ChArray ChString) connection [toChType $ map BS.singleton [0..255]]
+  -- runSerializationTest @(ChArray ChInt64) connection [toChType [0 :: ChInt64 .. 255]]
